Storing Service cost values as Int cents (minor unit)

+++ src/main/kotlin/com/thiagotoazza/data/models/services/ServiceRequest.kt
@@ -8,5 +8,5 @@ data class ServiceRequest(
     val vehicle: VehicleRequest,
     val date: Long,
     val typeId: String,
-    val cost: Double,
+    val cost: Int,
 )
